Fixed debug mode errors related to the audio system for cheeseboy, and seperated model functions into GLmesh.h/.cpp
