+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/GLd3d.h
@@ -15,6 +15,13 @@
 *** with this code. If not, see <http://www.gnu.org/licenses/>
 **/
 
+#include <string>
+
+namespace enigma {
+    void d3d_light_update_positions();
+}
+
+// ***** RENDER STATE MODE CONSTANTS *****
 enum {
   rs_fill,
   rs_line,
@@ -72,6 +79,7 @@ void d3d_set_fog_density(double density);
 void d3d_set_depth(double dep);
 void d3d_set_shading(bool smooth);
 
+// ***** PRIMITIVES BEGIN *****
 void d3d_primitive_begin(int kind);
 void d3d_primitive_begin_texture(int kind, int texId);
 void d3d_primitive_end();
@@ -87,6 +95,7 @@ void d3d_set_projection(double xfrom,double yfrom,double zfrom,double xto,double
 void d3d_set_projection_ext(double xfrom,double yfrom,double zfrom,double xto,double yto,double zto,double xup,double yup,double zup,double angle,double aspect,double znear,double zfar);
 void d3d_set_projection_ortho(double x, double y, double width, double height, double angle);
 void d3d_set_projection_perspective(double x, double y, double width, double height, double angle);
+// ***** PRIMITIVES END *****
 
 void d3d_draw_wall(double x1, double y1, double z1, double x2, double y2, double z2, int texId, int hrep, int vrep);
 void d3d_draw_floor(double x1, double y1, double z1, double x2, double y2, double z2, int texId, int hrep, int vrep);
@@ -94,9 +103,10 @@ void d3d_draw_block(double x1, double y1, double z1, double x2, double y2, doubl
 void d3d_draw_cylinder(double x1, double y1, double z1, double x2, double y2, double z2, int texid, int hrep, int vrep, bool closed, int steps);
 void d3d_draw_cone(double x1, double y1, double z1, double x2, double y2, double z2, int texid, double hrep, double vrep, bool closed, int steps);
 void d3d_draw_ellipsoid(double x1, double y1, double z1, double x2, double y2, double z2, int texid, double hrep, double vrep, int steps);
-void d3d_draw_icosahedron();
+void d3d_draw_icosahedron(int texid);
 void d3d_draw_torus(double x1, double y1, double z1, int texId, int hrep, int vrep, int csteps, int tsteps, double radius, double tradius, double TWOPI = 2*3.14);
 
+// ***** TRANSFORMATIONS BEGIN *****
 void d3d_transform_set_identity();
 void d3d_transform_add_translation(double xt,double yt,double zt);
 void d3d_transform_add_scaling(double xs,double ys,double zs);
@@ -116,7 +126,9 @@ void d3d_transform_stack_clear();
 bool d3d_transform_stack_empty();
 bool d3d_transform_stack_top();
 bool d3d_transform_stack_disgard();
+// ***** TRANSFORMATIONS END *****
 
+// ***** LIGHTS BEGIN *****
 bool d3d_light_define_direction(int id, double dx, double dy, double dz, int col);
 bool d3d_light_define_point(int id, double x, double y, double z, double range, int col);
 void d3d_light_define_specularity(int id, int r, int g, int b, double a);
@@ -124,37 +136,4 @@ void d3d_light_specularity(int facemode, int r, int g, int b, double a);
 void d3d_light_shininess(int facemode, int shine);
 void d3d_light_define_ambient(int col);
 bool d3d_light_enable(int id, bool enable);
-
-namespace enigma {
-    void d3d_light_update_positions();
-}
-
-#include <string>
-
-unsigned int d3d_model_create();
-void d3d_model_destroy(const unsigned int id);
-void d3d_model_copy(const unsigned int id, const unsigned int source);
-unsigned int d3d_model_duplicate(const unsigned int source);
-bool d3d_model_exists(const unsigned int id);
-void d3d_model_clear(const unsigned int id);
-void d3d_model_save(const unsigned int id, std::string fname);
-bool d3d_model_load(const unsigned int id, std::string fname);
-void d3d_model_draw(const unsigned int id, double x, double y, double z, int texId);
-void d3d_model_primitive_begin(const unsigned int id, int kind);
-void d3d_model_primitive_end(const unsigned int id);
-void d3d_model_vertex(const unsigned int id, double x, double y, double z);
-void d3d_model_vertex_color(const unsigned int id, double x, double y, double z, int col, double alpha);
-void d3d_model_vertex_texture(const unsigned int id, double x, double y, double z, double tx, double ty);
-void d3d_model_vertex_texture_color(const unsigned int id, double x, double y, double z, double tx, double ty, int col, double alpha);
-void d3d_model_vertex_normal(const unsigned int id, double x, double y, double z, double nx, double ny, double nz);
-void d3d_model_vertex_normal_color(const unsigned int id, double x, double y, double z, double nx, double ny, double nz, int col, double alpha);
-void d3d_model_vertex_normal_texture(const unsigned int id, double x, double y, double z, double nx, double ny, double nz, double tx, double ty);
-void d3d_model_vertex_normal_texture_color(const unsigned int id, double x, double y, double z, double nx, double ny, double nz, double tx, double ty, int col, double alpha);
-void d3d_model_wall(const unsigned int id, double x1, double y1, double z1, double x2, double y2, double z2, int hrep, int vrep);
-void d3d_model_floor(const unsigned int id, double x1, double y1, double z1, double x2, double y2, double z2, int hrep, int vrep);
-void d3d_model_block(const unsigned int id, double x1, double y1, double z1, double x2, double y2, double z2, int hrep, int vrep, bool closed = true);
-void d3d_model_cylinder(const unsigned int id, double x1, double y1, double z1, double x2, double y2, double z2, int hrep, int vrep, bool closed, int steps);
-void d3d_model_cone(const unsigned int id, double x1, double y1, double z1, double x2, double y2, double z2, int hrep, int vrep, bool closed, int steps);
-void d3d_model_ellipsoid(const unsigned int id, double x1, double y1, double z1, double x2, double y2, double z2, int hrep, int vrep, int steps);
-void d3d_model_icosahedron(const unsigned int id);
-void d3d_model_torus(const unsigned int id, double x1, double y1, double z1, int hrep, int vrep, int csteps, int tsteps, double radius, double tradius, double TWOPI = 2*3.14);
+// ***** LIGHTS END *****
