+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/include.h
@@ -3,7 +3,9 @@
 #include "GLbackground.h"
 #include "GLvertexbuffer.h"
 #include "GLshader.h"
+#include "GLtextures.h"
 #include "GLmaterial.h"
+#include "GLmesh.h"
 
 #include "GLfont.h"
 #include "GLcurves.h"
